Mostly remove the instructions on phpmyadmin (#123)
As I noted in issue #85, the section on the phpMyAdmin migration mainly repeats the procedures already set out, substituting the use of phpmyadmin for a database export and import into a pre-install CiviCRM db. The steps for the db export in this phpmyadmin section may or may not create errors, and importing a full db dump, including cache tables, into an empty db, is simpler and cleaner. 

The whole phpmyadmin section has been removed, and a brief note on using phpMyAdmin for export/import added in the main sequence of text. This section is still available publicly at https://wiki.civicrm.org/confluence/display/CRMDOC42/Alternate+Moving+the+database+using+phpMyAdmin