dev/wordpress#66 - Document transition for `wp-content/plugins/files/civicrm`
Discussion of the impetus: https://lab.civicrm.org/dev/wordpress/-/issues/66

Note: The text here refers to changes in 5.29.  The impetus, however, is
that the changes were first released in 5.27.  The expectation is that is to
revert from 5.27 and then release modified form in 5.29.  However, we should
have the docs ready hyperlinking in various comms.